Moving files nilearn.group_sparse_covariance to nilearn.connectome Initialized from nilearn.connectome Moved test files
