+++ nilearn/connectome/__init__.py
@@ -1,1 +1,9 @@
+"""
+Tools for computing functional connectivity matrices and also
+implementation of algorithm for sparse multi subjects learning
+of Gaussian graphical models.
+"""
+
 from .connectivity_matrices import sym_to_vec, ConnectivityMeasure
+
+from .group_sparse_covariance import GroupSparseCovariance, GroupSparseCovarianceCV
