+++ nilearn/connectome/tests/test_group_sparse_covariance.py
@@ -2,10 +2,9 @@ from nose.tools import assert_equal, assert_true, assert_raises
 
 import numpy as np
 from nilearn._utils.testing import generate_group_sparse_gaussian_graphs
-from nilearn.group_sparse_covariance import (group_sparse_covariance,
-                                             group_sparse_scores,
-                                             GroupSparseCovariance,
-                                             GroupSparseCovarianceCV)
+from nilearn.connectome.group_sparse_covariance import (group_sparse_covariance,
+                                                        group_sparse_scores)
+from nilearn.connectome import GroupSparseCovariance, GroupSparseCovarianceCV
 
 
 def test_group_sparse_covariance():
